+++ package.json
@@ -20,6 +20,7 @@
     "@fortawesome/free-solid-svg-icons": "^5.3.1",
     "@fortawesome/react-fontawesome": "^0.1.3",
     "polished": "^2.2.0",
+    "rollup-plugin-includepaths": "^0.2.3",
     "typeface-lato": "^0.0.54"
   },
   "devDependencies": {
@@ -99,6 +100,10 @@
   },
   "jest": {
     "collectCoverage": true,
+    "moduleDirectories": [
+      "node_modules",
+      "src"
+    ],
     "moduleFileExtensions": [
       "ts",
       "tsx",
