+++ src/Theme.story.tsx
@@ -2,8 +2,9 @@ import { storiesOf } from '@storybook/react';
 import { margin, size } from 'polished';
 import React, { DetailedHTMLProps, HTMLAttributes } from 'react';
 import { StyledComponentClass, withTheme } from 'styled-components';
-import Typography from '../src/atoms/Typography';
-import styled from './styled-components';
+
+import styled from './_styled-components';
+import { Typography } from './atoms';
 import Theme from './Theme';
 
 const Color = styled.div`
