+++ src/atoms/Button/Button.test.tsx
@@ -1,7 +1,8 @@
 import 'jest-dom/extend-expect';
 import React from 'react';
 import { fireEvent, render } from 'react-testing-library';
-import Button from '.';
+
+import Button from './Button';
 
 test('Button', () => {
   const handleClick = jest.fn();
