+++ src/atoms/Button/Button.tsx
@@ -1,9 +1,10 @@
 import { padding, transitions } from 'polished';
 import { ButtonHTMLAttributes, DetailedHTMLProps } from 'react';
 import { StyledComponentClass } from 'styled-components';
-import styled from '../../styled-components';
-import Theme, { borderRadius, scale, transitionDuration } from '../../Theme';
-import Typography from '../Typography';
+
+import styled from '_styled-components';
+import { Typography } from 'atoms';
+import Theme, { borderRadius, scale, transitionDuration } from 'Theme';
 
 interface Props {
   large?: boolean;
