+++ src/atoms/Heading/Heading.test.tsx
@@ -1,7 +1,8 @@
 import 'jest-dom/extend-expect';
 import React from 'react';
 import { render } from 'react-testing-library';
-import Heading from '.';
+
+import Heading from './Heading';
 
 test('Heading', () => {
   const { getByText, rerender } = render(<Heading>Heading</Heading>);
