+++ src/atoms/Heading/Heading.tsx
@@ -1,8 +1,9 @@
 import { DetailedHTMLProps, HTMLAttributes } from 'react';
 import { StyledComponentClass } from 'styled-components';
-import styled from '../../styled-components';
-import Theme, { scale } from '../../Theme';
-import Typography from '../Typography';
+
+import styled from '_styled-components';
+import { Typography } from 'atoms';
+import Theme, { scale } from 'Theme';
 
 export const Heading = styled(Typography)`
   color: ${(props: any) => props.as === 'h1' && props.theme.headline};
