+++ src/atoms/Icon/Icon.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { render } from 'react-testing-library';
-import Icon, { IconName, icons } from '.';
+
+import Icon, { IconName, icons } from './Icon';
 
 test('Icon', () => {
   for (const { iconName } of icons) {
