+++ src/atoms/Input/Input.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { render } from 'react-testing-library';
-import Input from '.';
+
+import Input from './Input';
 
 test('Input', () => {
   render(<Input />);
