+++ src/atoms/Input/Input.tsx
@@ -1,9 +1,10 @@
 import { padding } from 'polished';
 import { DetailedHTMLProps, InputHTMLAttributes } from 'react';
 import { StyledComponentClass } from 'styled-components';
-import styled from '../../styled-components';
-import Theme, { borderRadius, scale, transitionDuration } from '../../Theme';
-import Typography from '../Typography';
+
+import styled from '_styled-components';
+import { Typography } from 'atoms';
+import Theme, { borderRadius, scale, transitionDuration } from 'Theme';
 
 export const Input = styled(Typography)`
   background: ${props => props.theme.controlBackground};
