+++ src/atoms/Typography/Typography.tsx
@@ -1,5 +1,6 @@
 import 'typeface-lato';
-import styled from '../../styled-components';
+
+import styled from '_styled-components';
 
 export const Typography = styled.p`
   color: ${props => props.theme.text};
