+++ src/molecules/ActionPanel/ActionPanel.tsx
@@ -5,10 +5,10 @@ import React, {
   ReactNode,
 } from 'react';
 import { StyledComponentClass } from 'styled-components';
-import Panel from '../../atoms/Panel';
-import Typography from '../../atoms/Typography';
-import styled from '../../styled-components';
-import Theme, { scale } from '../../Theme';
+
+import styled from '_styled-components';
+import { Panel, Typography } from 'atoms';
+import Theme, { scale } from 'Theme';
 
 const ActionPanelBody = styled.div`
   ${padding(scale(1), scale(2))};
