+++ src/organisms/ComboBox/ComboBox.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { render } from 'react-testing-library';
-import ComboBox from '.';
+
+import ComboBox from './ComboBox';
 
 test('ComboBox', async () => {
   const { rerender } = render(<ComboBox items={new Set(['a', 'b', 'ab'])} />);
